requirements.txt
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
--extra-index-url https://download.pytorch.org/whl/cu124

faiss-cpu
sentence-transformers
scikit-learn
pandas
flask
kagglehub
python-dotenv
